+++ CHANGELOG.md
@@ -100,6 +100,9 @@ format.
 - Treat semantic static export link relations such as `rel="author"` as
   crawlable page links instead of copied assets.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
+- Rewrite common lazy-loading `data-*` asset attributes in static exports,
+  including background URLs and background `srcset` values on non-image tags.
+  ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
 
 ### Documentation